mcp_client/requirements.txt
streamlit==1.32.0
requests==2.31.0
sseclient-py==1.8.0
